move WORKSPACE+Builder
